# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Courses API",
        "description": "API REST da plataforma de cursos online AulasViolão. Autenticação via JWT Bearer Token.",
        "contact": {
            "name": "Big Space Creative",
            "email": "admin@digitalcourses.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desenvolvimento"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/api/v1/admin/dashboard": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Dashboard administrativo",
                "description": "Retorna métricas gerais: totais de usuários, cursos e matrículas. Exclusivo para admins.",
                "operationId": "adminDashboard",
                "responses": {
                    "200": {
                        "description": "Dashboard carregado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard carregado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "users": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer"
                                                        },
                                                        "students": {
                                                            "type": "integer"
                                                        },
                                                        "instructors": {
                                                            "type": "integer"
                                                        },
                                                        "admins": {
                                                            "type": "integer"
                                                        },
                                                        "premium": {
                                                            "type": "integer"
                                                        },
                                                        "free": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "courses": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer"
                                                        },
                                                        "published": {
                                                            "type": "integer"
                                                        },
                                                        "draft": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "enrollments": {
                                                    "properties": {
                                                        "total": {
                                                            "type": "integer"
                                                        },
                                                        "active": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso negado — somente admins"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Listar usuários",
                "description": "Lista todos os usuários com paginação. Suporta filtro por role e busca por nome/e-mail.",
                "operationId": "adminListUsers",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Filtrar por role",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "student",
                                "instructor",
                                "admin"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca por nome ou e-mail",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "alice"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página (padrão 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuários listados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuários listados com sucesso"
                                        },
                                        "data": {
                                            "description": "Resposta paginada do Laravel",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Exibir usuário",
                "description": "Retorna detalhes de um usuário incluindo seus cursos matriculados.",
                "operationId": "adminShowUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário encontrado"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso negado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token JWT obtido no login: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}